crud users datatable

+++ app/Services/User/UserService.php
@@ -4,42 +4,28 @@ namespace App\Services\User;
 
 use App\Models\User;
 use Illuminate\Contracts\View\View;
+use Illuminate\Support\Facades\Auth;
 
 class UserService
 {
     public function create(array $newData) : void
     {
-        $templatePassword = '12345678';
-        $newData['password'] = $templatePassword;
-
         User::create($newData);
     }
 
-    public function update(array $editedData, User $user) : void
+    public function update(array $editedData, User $user): void
     {
+        if (array_key_exists('password', $editedData) && $editedData['password'] === null) {
+            unset($editedData['password']);
+        }
+
         $user->update($editedData);
     }
 
     public function delete(User $user) : void
     {
-        $user->delete();
-    }
-
-    public function indexViewInCreateMode(): View
-    {
-        $users = User::all();
-        $inCreateMode = true;
-        $roles = ['admin', 'user'];
-
-        return view('admin.users.index', compact('users', 'inCreateMode', 'roles'));
-    }
-
-    public function indexViewInEditMode(User $user): View
-    {
-        $users = User::all();
-        $editUserId = $user->id;
-        $roles = ['admin', 'user'];
-
-        return view('admin.users.index', compact('users', 'editUserId', 'roles'));
+        if (Auth::id() !== $user->id) {
+            $user->delete();
+        }
     }
 }
